fix: Improve transparency algorithm
It's still not reliable, but it will regularly get close neough to the edge that it works.

I think I need a "gradient" pass; the angle of the line from the previous point to this point can't exceed 20º of the previous angle. This would allow for rounded corners, but not random blips into the page.