feat: :sparkles: 3D models that represent postcard transparency
This (large and broadly untested 😅) change introduces "drawing around" the transparency at the edges of a postcard to make a 3D model which resembles a postcard image with transparency.

It introuces the ability to ignore transparency when encoding for the web codec (so that the USDs can use a smaller JPEG texture, instead of a PNG). There's a weird bug I haven't pinned down where very complicated edges (eg. Seattle) break at low values for ep(silon) on the simplifier.

I think it's because the macOS USD renderer can't cope with insanely large faces — but that doesn't *quite* ring true.